+++ alex/applications/PublicTransportInfoCS/wshub_lz.sh
@@ -2,7 +2,7 @@
 
 pushd ..
 
-./wshub.py -c ./PublicTransportInfoCS/ptics.cfg ../resources/private/ext-lz-277278190.cfg ./PublicTransportInfoCS/kaldi.cfg ./PublicTransportInfoCS/ptics_hdc_slu.cfg
+./wshub.py -c ./PublicTransportInfoCS/ptics.cfg ./PublicTransportInfoCS/kaldi.cfg ./PublicTransportInfoCS/ptics_hdc_slu.cfg
 
 
 popd
