+++ alex/applications/PublicTransportInfoCS/wshub_metacentrum
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+pushd ..
+
+./wshub.py -c ./PublicTransportInfoCS/ptics.cfg ./PublicTransportInfoCS/kaldi.cfg ./PublicTransportInfoCS/ptics_hdc_slu.cfg ./PublicTransportInfoCS/wshub_metacentrum.cfg
+
+
+popd
