+++ alex/applications/PublicTransportInfoCS/wshub_metacentrum.cfg
@@ -1,0 +1,29 @@
+#!/usr/bin/env python
+# vim: set fileencoding=UTF-8 filetype=python :
+#
+#  When the configuration file is loaded, several automatic transformations
+#  are applied:
+#
+# 1) '{cfg_abs_path}' as a substring of atomic attributes is replaced by
+#    an absolute path of the configuration files.  This can be used to
+#    make the configuration file independent of the location of programs
+#    using the configuration file.
+#
+# or better user use the as_project_path function
+
+from __future__ import unicode_literals
+
+from alex.utils.config import as_project_path
+
+from alex.applications.PublicTransportInfoCS.hdc_slu import PTICSHDCSLU
+from alex.applications.PublicTransportInfoCS.preprocessing import PTICSSLUPreprocessing
+
+config = {
+  'WSIO': {
+        'router_addr': '147.251.253.69',
+        'router_port': 5000,
+        'listen_addr': '0.0.0.0',
+        'listen_port': 8000,
+        'alex_addr': 'ws://147.251.253.69:8000/'
+  }
+}
