+++ alex/resources/default.cfg
@@ -256,6 +256,20 @@ config = {
         'blacklist_for': 2 * 60 * 60,            # in seconds
         'limit_reached_message': u'Thank you for calling. Your calling limit was reached. Please call later.',
     },
+    'WSHub': {
+        'pid_file': as_project_path("applications/wshub.pid"),
+        'wait_time_before_calling_back': 10,
+        'history_length': 1000,
+        'hard_time_limit': 6 * 60,  # maximal length of a dialogue in seconds
+        'hard_turn_limit': 120,   # maximal number of turn in a dialogue
+
+        'call_db': './call_logs/call_db.pckl',
+        'period': 48 * 60 * 60,    # in seconds
+        'last_period_max_num_calls': 200,
+        'last_period_max_total_time': 120 * 60,  # in seconds
+        'blacklist_for': 2 * 60 * 60,            # in seconds
+        'limit_reached_message': u'Thank you for calling. Your calling limit was reached. Please call later.',
+    },
     'WebHub': {
         'port': 8000,
     },
