+++ docker/install_dep.sh
@@ -25,4 +25,4 @@ make install
 cd /app/pjsip/pjsip-apps/src/python/
 python setup-pjsuaxt.py install
 
-apt-get install python-speex
+apt-get install libspeex-dev libspeexdsp-dev
